requirements.txt
fpdf==1.7.2
python-dotenv~=0.15.0